requirements.txt
# FastAPI and related
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]==2.6.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
neo4j==5.14.1
python-dotenv==1.0.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1

# Storage
minio==7.2.0

# Web Automation
playwright==1.40.0

# PDF Generation
weasyprint==60.2
jinja2==3.1.3

# AI and ML
openai>=1.10.0,<2.0.0
langchain==0.1.4
langchain-openai==0.0.5

# Utilities
python-dateutil==2.8.2
uuid==1.30
python-docx==0.8.11

# Testing
pytest==7.4.3
httpx==0.26.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Code Quality
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Additional dependencies
pdfminer.six==20221105
aiofiles==23.2.1 